+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ---
 
+#### 21.06.2017
+
+* [Java] Grade School (Exercism)
+
 #### 20.06.2017
 
 * [Java] Linked List (Exercism)